// src/app/models/profile/compute-profile-response.ts
export interface ComputeProfileResponse {
  age: number;
  gender: string;
  height: number;
  weight: number;
  bmr: number;
  activityFactor: number;
  tdee: number;
  fitnessGoalId: string;
  calorieAdjustment: number;
  targetCalories: number;
  targetProtein: number;
  targetFat: number;
  targetCarbs: number;
  activityLevelName: string;
  goalName: string;
}